@@ -1,12 +1,14 @@
 cmake_minimum_required(VERSION 3.10)
 
 # C++ 표준 설정
-set(CMAKE_CXX_STANDARD 11)
+set(CMAKE_CXX_STANDARD 17)
 set(CMAKE_CXX_STANDARD_REQUIRED True)
 
+message("CMAKE in Component c++ 17")
+
 set (SOURCES
-#     ./src/Activity.cpp
-#     ./src/Facade.cpp
+     ./src/Activity.cpp
+     ./src/Facade.cpp
      ./src/Map.cpp
      ./src/Unit.cpp
      ./src/User.cpp)
