@@ -1,8 +1,10 @@
 cmake_minimum_required(VERSION 3.10)
 
 # C++ 표준 설정
-set(CMAKE_CXX_STANDARD 11)
+set(CMAKE_CXX_STANDARD 17)
 set(CMAKE_CXX_STANDARD_REQUIRED True)
+
+message("CMAKE in Factory c++ 17")
 
 set (SOURCES
      ./src/Factory.cpp)
